refactor: (#1) Detekt Rules Violeted Code Change
