+++ src/main/kotlin/com/dsm/plugins/HTTP.kt
@@ -1,8 +1,10 @@
-package com.dsm_delivery.plugins
+package com.dsm.plugins
 
-import io.ktor.http.*
-import io.ktor.server.plugins.cors.routing.*
-import io.ktor.server.application.*
+import io.ktor.http.HttpHeaders
+import io.ktor.http.HttpMethod
+import io.ktor.server.application.Application
+import io.ktor.server.application.install
+import io.ktor.server.plugins.cors.routing.CORS
 
 /**
  *
@@ -19,6 +21,6 @@ fun Application.configureHTTP() {
         allowMethod(HttpMethod.Delete)
         allowMethod(HttpMethod.Patch)
         allowHeader(HttpHeaders.Authorization)
-        anyHost() // @TODO: Don't do this in production if possible. Try to limit it.
+        anyHost() // @TODO Don't do this in production if possible. Try to limit it.
     }
 }
