+++ src/main/kotlin/com/dsm/plugins/Monitoring.kt
@@ -1,9 +1,10 @@
-package com.dsm_delivery.plugins
+package com.dsm.plugins
 
-import io.ktor.server.plugins.callloging.*
-import org.slf4j.event.*
-import io.ktor.server.request.*
-import io.ktor.server.application.*
+import io.ktor.server.application.Application
+import io.ktor.server.application.install
+import io.ktor.server.plugins.callloging.CallLogging
+import io.ktor.server.request.path
+import org.slf4j.event.Level
 
 /**
  *
