+++ src/main/kotlin/com/dsm/plugins/Security.kt
@@ -1,15 +1,16 @@
-package com.dsm_delivery.plugins
+package com.dsm.plugins
 
 import com.auth0.jwt.JWT
 import com.auth0.jwt.algorithms.Algorithm
-import com.dsm_delivery.domain.auth.token.JwtGenerator
-import com.dsm_delivery.persistence.repository.StudentRepository
-import io.ktor.http.*
-import io.ktor.server.application.*
-import io.ktor.server.auth.*
-import io.ktor.server.auth.jwt.*
-import io.ktor.server.config.*
-import io.ktor.server.response.*
+import com.dsm.domain.auth.token.JwtGenerator
+import com.dsm.persistence.repository.StudentRepository
+import io.ktor.http.HttpStatusCode
+import io.ktor.server.application.Application
+import io.ktor.server.auth.authentication
+import io.ktor.server.auth.jwt.JWTPrincipal
+import io.ktor.server.auth.jwt.jwt
+import io.ktor.server.config.ApplicationConfig
+import io.ktor.server.response.respond
 import org.koin.ktor.ext.getKoin
 import java.util.UUID
 
@@ -20,28 +21,7 @@ import java.util.UUID
  * @author Chokyunghyeon
  * @date 2023/03/16
  **/
-class SecurityProperties(config: ApplicationConfig) {
-
-    val realm: String = config.property(JWT_REALM).getString()
-    val secret: String = config.property(JWT_SECRET).getString()
-    val audience: String = config.property(JWT_AUDIENCE).getString()
-    val issuer: String = config.property(JWT_ISSUER).getString()
-
-    val refreshExpired: Long = config.property(REFRESH_TOKEN_EXPIRED_TIME).getString().toLong()
-    val accessExpired: Long = config.property(ACCESS_TOKEN_EXPIRED_TIME).getString().toLong()
-
-    private companion object {
-        const val JWT_AUDIENCE: String = "jwt.audience"
-        const val JWT_SECRET: String = "jwt.secret"
-        const val JWT_REALM: String = "jwt.realm"
-        const val JWT_ISSUER: String = "jwt.issuer"
-        const val REFRESH_TOKEN_EXPIRED_TIME = "jwt.token.refresh-expired"
-        const val ACCESS_TOKEN_EXPIRED_TIME = "jwt.token.access-expired"
-    }
-}
-
 fun Application.configureSecurity() {
-
     val securityProperties: SecurityProperties = getKoin().get()
     val studentRepository: StudentRepository = getKoin().get()
 
@@ -72,5 +52,25 @@ fun Application.configureSecurity() {
                 call.respond(HttpStatusCode.Unauthorized, "Token is not valid or has expired")
             }
         }
+    }
+}
+
+class SecurityProperties(config: ApplicationConfig) {
+
+    val realm: String = config.property(JWT_REALM).getString()
+    val secret: String = config.property(JWT_SECRET).getString()
+    val audience: String = config.property(JWT_AUDIENCE).getString()
+    val issuer: String = config.property(JWT_ISSUER).getString()
+
+    val refreshExpired: Long = config.property(REFRESH_TOKEN_EXPIRED_TIME).getString().toLong()
+    val accessExpired: Long = config.property(ACCESS_TOKEN_EXPIRED_TIME).getString().toLong()
+
+    private companion object {
+        const val JWT_AUDIENCE: String = "jwt.audience"
+        const val JWT_SECRET: String = "jwt.secret"
+        const val JWT_REALM: String = "jwt.realm"
+        const val JWT_ISSUER: String = "jwt.issuer"
+        const val REFRESH_TOKEN_EXPIRED_TIME = "jwt.token.refresh-expired"
+        const val ACCESS_TOKEN_EXPIRED_TIME = "jwt.token.access-expired"
     }
 }
